Move stores list in StoreSelect to separate component
