+++ dashboard/src/screens/StoreSelect.tsx
@@ -1,30 +1,11 @@
 import React from 'react';
-import {
-	View,
-	Text,
-	TouchableOpacity,
-	FlatList,
-	StyleSheet
-} from 'react-native';
-import { useDispatch } from 'react-redux';
+import { Text, TouchableOpacity, StyleSheet } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { updatePreference } from '../redux/preferences/actions';
-import { useManagedStoresQuery } from '../types/api';
-import { useAppSelector } from '../redux/store';
+import StoresList from '../components/store-select/StoresList';
 
 const StoreSelect: React.FC = () => {
-	const dispatch = useDispatch();
 	const { navigate } = useNavigation();
-	const userId = useAppSelector(({ auth }) => auth.userId);
-	const [{ data }] = useManagedStoresQuery({ variables: { userId } });
-
-	const managers = data?.store_managers;
-	const stores = managers?.map(({ store }) => store);
-
-	const handleStoreSelect = (storeId: string) => {
-		dispatch(updatePreference({ activeStore: storeId }));
-	};
 
 	return (
 		<SafeAreaView style={styles.container}>
@@ -32,15 +13,7 @@ const StoreSelect: React.FC = () => {
 			<Text style={{ fontSize: 17 }}>
 				Kindly select a store you manage, or create a new one:
 			</Text>
-			<FlatList
-				horizontal
-				data={stores}
-				renderItem={({ item }) => (
-					<TouchableOpacity onPress={() => handleStoreSelect(item.id)}>
-						<Text>{item.name}</Text>
-					</TouchableOpacity>
-				)}
-			/>
+			<StoresList />
 			<TouchableOpacity
 				onPress={() => navigate('CreateStore')}
 				style={{ alignSelf: 'center', marginBottom: 16 }}
